+++ src/scripts/flight/tracking/map.ts
@@ -1,0 +1,1 @@
+export {}
