+++ src/scripts/flight/tracking/trackingConstants.ts
@@ -1,0 +1,2 @@
+export const defaultTrackingInterval = 5000
+export const defaultTrackingDecimal = 4
